# backend/src/main/java/com/spectemus/simul/backend/telemetry/TelemetryEventType.java
package com.spectemus.simul.backend.telemetry;

public enum TelemetryEventType {
    FIRST_FRAME,
    PLAYBACK_ERROR,
    PUBLISH_START,
    PUBLISH_FAILURE,
    QUALITY_SUMMARY,
    RECOVERY_REQUESTED,
    RECOVERY_STARTED,
    RECOVERY_SUCCEEDED,
    RECOVERY_FAILURE
}