# backend/src/main/java/com/spectemus/simul/backend/telemetry/TelemetryRequest.java
package com.spectemus.simul.backend.telemetry;

import java.util.List;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;

public record TelemetryRequest(@NotEmpty @Size(max = 50) @Valid List<TelemetryEvent> events) {}