# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Spectemus Simul REST API
  version: 1.0.0
  description: |
    Контракт product backend. Эндпоинты с `x-implementation-status: planned`
    фиксируют согласованную будущую API surface и не считаются реализованными до
    соответствующего тикета.
servers:
  - url: /
    description: Same-origin API через development или production reverse proxy
...
paths:
...
  /api/v1/telemetry:
    post:
      tags: [Telemetry]
      operationId: submitTelemetry
      summary: Принять батч privacy-safe client telemetry
      description: |
        Endpoint принимает батч client telemetry событий (first frame, publish/playback
        ошибки, quality summary) для агрегированных beta-метрик. Он не принимает медиафайлы,
        chat history, room secrets, LiveKit tokens или identity данные. Каждое событие даёт
        только aggregate counter и low-cardinality tag; roomId, role и detail используются
        лишь для log-корреляции и никогда не становятся метрик-тегами.
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryRequest"
      responses:
        "202":
          description: Батч принят и получил correlation-friendly receipt
          headers:
            Cache-Control:
              description: Ответ содержит диагностический receipt и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wt_session
      description: |
        Opaque session credential. Production cookie: HttpOnly, Secure, SameSite=Strict.
    accountSessionCookie:
      type: apiKey
      in: cookie
      name: wt_account
      description: |
        Planned opaque account credential for Internet mode. Не заменяет
        `wt_session` и не передаётся в URL, JavaScript storage или WebSocket query.
    feedbackAdminToken:
      type: apiKey
      in: header
      name: X-Feedback-Admin-Token
      description: Operator token from `FEEDBACK_ADMIN_TOKEN`; disabled when env is empty.
...
  schemas:
...
    TelemetryEventType:
      type: string
      enum:
        - FIRST_FRAME
        - PLAYBACK_ERROR
        - PUBLISH_START
        - PUBLISH_FAILURE
        - QUALITY_SUMMARY
        - RECOVERY_REQUESTED
        - RECOVERY_STARTED
        - RECOVERY_SUCCEEDED
        - RECOVERY_FAILURE
    TelemetryQualityStatus:
      type: string
      enum: [GOOD, WARNING, POOR, LOST, UNKNOWN]
    TelemetryEvent:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          $ref: "#/components/schemas/TelemetryEventType"
        roomId:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        role:
          type: string
          enum: [HOST, GUEST]
        qualityStatus:
          $ref: "#/components/schemas/TelemetryQualityStatus"
        detail:
          type: string
          maxLength: 200
    TelemetryRequest:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/TelemetryEvent"
    TelemetryResponse:
      type: object
      additionalProperties: false
      required: [telemetryId, correlationId, receivedAt, accepted]
      properties:
        telemetryId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        receivedAt:
          type: string
          format: date-time
        accepted:
          type: integer
          minimum: 0
...
    ProblemDetails:
      $ref: "./schemas/problem-details.schema.json"
  responses:
    BadRequest:
      description: Некорректный JSON или формат запроса
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ValidationFailed:
      description: Запрос не прошёл предметную валидацию
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Внутренняя ошибка без раскрытия реализации
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"